rename(ui-design-tool): add index file of hooks
