+++ packages/ui-design-tool/hooks/index.ts
@@ -1,0 +1,2 @@
+export { default as useUIRecord } from './useUIRecord';
+export { default as useUIRecordForInteraction } from './useUIRecordForInteraction';
